app/backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.10

# Cache and Message Queue
redis==5.0.1
django-redis==5.4.0
celery==5.3.4

# CORS and Filters
django-cors-headers==4.3.1
django-filter==24.2

# API Documentation
drf-spectacular==0.26.5

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# Development and Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Utilities
python-decouple==3.8
Pillow==11.0.0
gunicorn==21.2.0
channels==4.0.0
channels-redis==4.1.0
dj-database-url==2.1.0
python-magic==0.4.27
PyYAML==6.0.2

# Algorithms dependencies
numpy>=1.21.0

# Celery dependencies
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13

# Export dependencies
openpyxl==3.1.5
reportlab==4.2.2
